backend/old_backend/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
openai>=1.3.0
asyncio
pathlib
urllib3>=2.0.0
lxml>=4.9.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
python-dateutil>=2.8.0
PyYAML>=6.0.0